fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pypandoc==1.12
python-docx==1.1.0
pandas==2.1.3
Pillow==10.1.0
aiofiles==23.2.1
